lab5/pyproject.toml
[tool.poetry]
name = "lab5"
version = "0.1.0"
description = ""
authors = ["Ivan Krasnopolsky <i1.krasnopolskiy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2 = "^2.9.9"
sqlalchemy = "^2.0.30"
pillow = "^10.3.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = '(\.git|\.cache|\.venv|\.idea|__pycache__)'
line-length = 120

[tool.isort]
line_length = 120
profile = "black"
skip = ["**/__pycache__/**", ".cache", ".git", ".idea", ".venv"]
